PR 496: Move build_sub_pipeline to execution.py
