+++ python_modules/dagster/dagster/utils/test.py
@@ -15,6 +15,7 @@ from dagster import (
     execute_pipeline,
 )
 
+from dagster.core.execution import build_sub_pipeline
 from dagster.core.execution_context import RuntimeExecutionContext
 
 
@@ -58,43 +59,6 @@ def get_temp_file_names(number):
 
 def _dep_key_of(solid):
     return SolidInstance(solid.definition.name, solid.name)
-
-
-def build_sub_pipeline(pipeline_def, solid_names):
-    '''
-    Build a pipeline which is a subset of another pipeline.
-    Only includes the solids which are in solid_names.
-    '''
-
-    check.inst_param(pipeline_def, 'pipeline_def', PipelineDefinition)
-    check.list_param(solid_names, 'solid_names', of_type=str)
-
-    solid_name_set = set(solid_names)
-    solids = list(map(pipeline_def.solid_named, solid_names))
-    deps = {_dep_key_of(solid): {} for solid in solids}
-
-    def _out_handle_of_inp(input_handle):
-        if pipeline_def.dependency_structure.has_dep(input_handle):
-            output_handle = pipeline_def.dependency_structure.get_dep(input_handle)
-            if output_handle.solid.name in solid_name_set:
-                return output_handle
-        return None
-
-    for solid in solids:
-        for input_handle in solid.input_handles():
-            output_handle = _out_handle_of_inp(input_handle)
-            if output_handle:
-                deps[_dep_key_of(solid)][input_handle.input_def.name] = DependencyDefinition(
-                    solid=output_handle.solid.name,
-                    output=output_handle.output_def.name,
-                )
-
-    return PipelineDefinition(
-        name=pipeline_def.name,
-        solids=list(set([solid.definition for solid in solids])),
-        context_definitions=pipeline_def.context_definitions,
-        dependencies=deps,
-    )
 
 
 def build_pipeline_with_input_stubs(pipeline_def, inputs):
